Introduction to C
This lab was used to introduce students to C, in a previous class we were introduced to C++ and had to adjust. Part of the adjustment was having to learn how to use the Linux terminal to compile and journey through your files. After getting comfortable we created simple programs that will display output to the console, read input from the user, use if statements and for loops in C, and perform some simple error handling.